commit6

+++ ft_memmove.c
@@ -8,7 +8,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 		return (0);
 	if (dst > src)
 	{
-		i = len - 1;
+		i = (int)len - 1;
 		while (i >= 0)
 		{
 			*(char *)(dst + i) = *(char *)(src + i);
@@ -18,7 +18,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	else
 	{
 		i = 0;
-		while (i < len)
+		while (i < (int)len)
 		{
 			*(char *)(dst + i) = *(char *)(src + i);
 			i ++;
